requirements.txt
appnope==0.1.4
asttokens==2.4.1
backoff==2.2.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
comm==0.2.1
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
executing==2.0.1
googleapis-common-protos==1.62.0
grpcio==1.60.1
honeycomb-opentelemetry==0.3.0a0
idna==3.6
importlib-metadata==6.11.0
ipykernel==6.29.2
ipython==8.21.0
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
packaging==23.2
parso==0.8.3
pexpect==4.9.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.8.2
pyzmq==25.1.2
requests==2.31.0
setuptools==69.1.0
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
urllib3==2.2.0
wcwidth==0.2.13
wrapt==1.16.0
zipp==3.17.0
